# Performance Testing of Fluent-bit with several filters shows log processing falling < 5mb/s