@ryanohnemus 2024-09-20
hi @agup006, 

**_Heya is the file already present prior to read? I’m wondering if iops are a contributing factor to slow down_**
> No the file is not present before fluent-bit starts. The file writing process handled by the sidecar container is generally complete within 8-10 seconds and should be realistic to how fluent-bit works within kubernetes. Typically there will be a daemonset of fluentbit that is reading container log files written by the container runtime (such as `containerd`) and those will all be generally using the same underlying disk.

**_multiple filters chained together is one we know is not optimal which is why we’ve introduced processors. Have you tried the benchmark with processors?_**
> I tried the setup with processors yesterday, this had 2 improvements: 
1) inputs were able to generally keep up with log rotations without issue, but only after switching to utilize filesystem storage. Memory storage still allows the input to get backed up and pause. 
2) ability to use multiple output workers, however the performance improvement here is only a few mb/s and from what I can tell i am hitting the same issue as #8245, which looks acknowledged but then it closed (not sure if it just went stale) but there is no performance improvement past 2-3 output workers regardless of how many cpu resources you allocate fluentbit

Drawbacks: 
1) you entirely lose filter/processor pipeline visibility. I no longer saw metrics from any filters that were defined as processors.

side note: 
- I also noticed the multiple filters bottleneck. For my setup I need to use a lua script, so I tried to combine my modifies (field rename, drop, and copy) from the `filters-extended` config directly into lua and saw an increase in mb/s because there were less filters. However the use of lua alone was seemingly a big bottleneck, is this expected?

**_Checking the commit also looks like latest? Thought I would doublecheck though._** 
> yes, this perf script is added on top of master but uses the latest helm chart which has an actual fluent-bit release it tests from. I can clean up the shellcheck warning if we end up wanting to merge this in.