@lecaros 2024-09-23
Hi there. Thanks for this PR.
I've run in and noted the performance you mention.
* what's your expectation for throughput?
* is filesystem storage out of the picture for any reason?
* I noticed you are handling the file rotation instead of leaving it to k8s runtime. Is there any reason for this?