@uristernik 2024-10-07
@ryanohnemus super interesting stuff!

Did you take into account CPU / Memory consumption when testing the different setups? We noticed that higher buffers cause very high memory utilization when having to deal with lots of logs