@ryanohnemus 2024-10-08
@uristernik I wasn't as concerned with the memory utilization for my testing, but the test could be repeated looking specifically at that. As far as CPU, I was pushing as much log data as I could through fluent-bit so I was expecting CPU utilization quite high. The main purpose of my test was for throughput, but if you're looking for ways to limit memory utilization there's a few things you can try: 

1) in your tail input plugin, use a mem_buf_limit, this will prevent fluent-bit from ingesting more records if it has at least that much data waiting to be flushed. This only works if you're not using `storage.type filesystem`. Keep in mind if your input is queueing up, you have a high probability of missing log rotations and therefore logs.

2) You can toy around with storage.max_chunks_up and lower it from the 128 default, this will tell the engine to use less chunks in memory. Chunks can vary in size, but are typically 2mb. If you are using a higher buffer size/chunk size on the tail input, you will go over the 2mb size pretty easily. This size >2mb is something that could be addressed by #9385.

3) If you're using processors (basically filter pipelines handled in the output thread), this will more than double your initial memory requirement (usage generally becomes num worker threads + 1 * previous memory limit). Processors run after the input chunk is handed off to an output thread. The original input chunk remains while each worker thread runs processors that can enhance/add records to the processed chunk that is finally delivered to the output plugin's flush callback.