@ryanohnemus 2025-01-11
Commenting so it doesn't go stale. This is probably still a good test to merge. It makes it easier for perf testing within a kubernetes cluster.